AutoTimer: Use a default recording destination of "None".
A "dirname" of None in a timer means that the location will be the
default recording location at the moment the recording starts.

Immediately filling in the value results in the recording forcibly
being put there, which is not what a user would expect if he did not
set a location. If he changed the location after the timer has been
programmed, the recording would still go to the old location.