AutoTimerConfiguration: Stop writing corrupt XML comments
Autotimer writes out service names in some incompatible encoding, resulting in a message
that the autotimer.xml file is corrupt and cannot be loaded.
The "corruption" appears to be invalid encoding of special characters in service names. This
occurs in the "comments" following the service ref. Since the comment isn't critical, just
stop saving the comment text to fix the issue.