transcodingsetup: fix config defaults
TranscodingSetup.__init__ does not run until the plugin is opened for
the first time.
Until then, the settings did not get their default values.

As a result, openwebif would crash when streaming was attempted before
the transcodingsetup plugin has been used for the first time, because it
tries to use some of the config.plugin.transcodingsetup values.

And since the defaults of these values are fixed, we might as well
initialize them in the config constructors.
(Only the port setting, which has to be runtime detected, is still
initialized in the TranscodingSetup.__init__)