transcodingsetup: avoid creating empty streamproxy.conf
When streamproxy.conf was missing, an empty streamproxy.conf file was
created when transcoding settings were modified.
No longer create an empty config file in that case.

Also, on hardware where the streamproxy is not required, do not show a
warning when streamproxy.conf is missing.
(The warning is still shown when streamproxy.conf is missing on hardware
where the streamproxy is a requirement)