networkbrowser: Allow to use host name for mounting
If your network is slightly dynamic, host names will be much more useful
than IP addresses for mounting network drives.
In order not to break things, add a "host" entry which is just a string
and use that instead of the IP if provided. This allows to mount say
mybox.local which will neatly resolve to the proper box.
In future we should integrate this with avahi, so you can just pick the
host from a list (and make the horribly slow "browser" obsolete).