blinscan: save MIS/PLS in xml just like we do in satellites.xml
This commit will save the MIS/PLS TP information in the same way
we are saving on satellites.xml. If TP is only MIS, only is_id will
be saved. If TP is only PLS, only pls_code and mode will be saved.
If TP is both PLS and MIS both info will be saved.