[AutoTimer] fix logic checkSimilarity
1)   At this point(Short Descriptions) we assume the similarity match to
be True unless we have been asked to check Extended Descriptions and we
have *both* Extended Descriptions in place; in which case we test them.
2)Remove shortdesc1 in shortdesc2
This is not true in principle and breaks all logic ratio.
E.g.
shortdesc1 = "news"
shortdesc2 = "Sunday news (weekly review)"
shortdesc1 in shortdesc2 --> answer True